tag: allow idempotent "git tag" without "--force"
When "git tag T O" is told to create a tag pointing at an object O
without the "--force" option, it refuses with "tag T already exists",
even when T points at O (which makes it a no-op).

Let's allow this "idempotent" case by special casing and making it
truly a no-op.

Signed-off-by: Junio C Hamano <gitster@pobox.com>