config.mak.uname: update settings for FreeBSD
Even though FreeBSD 6 introduced memmem(), the implementation in
that version was buggy and not performant until FreeBSD 12, FreeBSD
11.4 (the last version of FreeBSD 11) went end of life in September
2021, so nobody should be using it since it hasn't had security
support since then.  And memmem() has even been functional (but
slow) since FreeBSD 11.0, and 10.4 went EOL in 2018.  So users
shouldn't actually be experiencing any actual functionality problems
since then.

Let's draw the line to require FreeBSD 12 or newer (but we do not
officially document it or enforce it by breaking build when compiled
on older versions, at least not yet), which allows us to drop the
special casing of FreeBSD 4.x and rely on platform implementation of
memmem() unconditionally.

Signed-off-by: Brad Smith <brad@comstyle.com>
[jc: log message with help from info by brian carlson]
Reviewed-by: Collin Funk <collin.funk1@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>